feat: Add Link import to Contact page
This commit adds the import statement for the Link component from the 'next/link' package to the Contact page. This import is necessary for using the Link component to create navigation links within the application. By including this import, we ensure that the Contact page can utilize the Link component for seamless navigation.