chore: Update MusicPlayer.module.css to fix formatting issue
The code changes in this commit fix a formatting issue in the MusicPlayer.module.css file. There was an extra newline at the end of the file, which has been removed. This change ensures that the file adheres to the coding style guidelines and improves the maintainability of the codebase.